backend/foodgram_django/api/serializers.py
from rest_framework import serializers

from recipes.models import Recipe
from users.models import User


class UserSerializer(serializers.ModelSerializer):
    password = serializers.HiddenField(
        default=serializers.CurrentUserDefault()
    )

    class Meta:
        model = User
        fields = (
            'email',
            'pk',
            'username',
            'first_name',
            'last_name',
            'password'
        )


class RecipeSerializer(serializers.ModelSerializer):
    class Meta:
        model = Recipe
        fields = (
            'author',
            'pk',
            'name',
            'text',
        )
